Revert "Merge branch 'edit/mdParser' into add/FindMsgByPR"
This reverts commit 432dd64adaba33a627e9e62c0abc7987cc2a5a65, reversing
changes made to 413e417c1fc80911ef4508b08aa85ab628c365ed.
