+++ gitcord/embeds.go
@@ -3,12 +3,31 @@ package gitcord
 import (
 	"fmt"
 	"strconv"
+	"strings"
 
 	"github.com/diamondburned/arikawa/v3/discord"
+	"github.com/diamondburned/ningen/v3/discordmd"
 	"github.com/ethanthatonekid/gitcord/gitcord/internal/discordclient"
-	"github.com/ethanthatonekid/gitcord/gitcord/internal/markdown"
 	"github.com/google/go-github/v47/github"
+	"github.com/yuin/goldmark/parser"
+	"github.com/yuin/goldmark/text"
 )
+
+type link struct {
+	title string
+	href  string
+}
+
+type links []link
+
+func (l links) commaSeparatedList() string {
+	var b []string
+	for _, link := range l {
+		b = append(b, convertHyperlink(link.title, link.href))
+	}
+
+	return strings.Join(b, ", ")
+}
 
 /// START IssuesEvent Discord embeds
 
@@ -23,21 +42,21 @@ func (c *Config) makeIssueEmbed(issue *github.Issue) discord.Embed {
 	if len(issue.Labels) > 0 {
 		fields = append(fields, discord.EmbedField{
 			Name:  "Labels",
-			Value: markdown.ConvertLabels(issue.Labels),
+			Value: convertLabels(issue.Labels),
 		})
 	}
 
 	if len(issue.Assignees) > 0 {
 		fields = append(fields, discord.EmbedField{
 			Name:  "Assignees",
-			Value: markdown.ConvertUsers(issue.Assignees),
+			Value: convertUsers(issue.Assignees),
 		})
 	}
 
 	if issue.Milestone != nil {
 		fields = append(fields, discord.EmbedField{
 			Name:  "Milestone",
-			Value: markdown.ConvertHyperlink(issue.Milestone.GetTitle(), issue.Milestone.GetHTMLURL()),
+			Value: convertHyperlink(issue.Milestone.GetTitle(), issue.Milestone.GetHTMLURL()),
 		})
 	}
 
@@ -56,7 +75,7 @@ func (c *Config) makeIssueEmbed(issue *github.Issue) discord.Embed {
 			Name: issue.GetUser().GetLogin(),
 			Icon: issue.GetUser().GetAvatarURL(),
 		},
-		Description: markdown.Convert(issue.GetBody(), issue.GetHTMLURL()),
+		Description: convertMarkdown(issue.GetBody(), issue.GetHTMLURL()),
 		Color:       c.ColorScheme.Color(IssueOpened, true),
 		Fields:      fields,
 	}
@@ -233,7 +252,7 @@ func (c *Config) makeIssueCommentEmbed(ev *github.IssueCommentEvent) discord.Emb
 
 	return discord.Embed{
 		Title:       title,
-		Description: markdown.Convert(comment.GetBody(), comment.GetHTMLURL()),
+		Description: convertMarkdown(comment.GetBody(), comment.GetHTMLURL()),
 		URL:         comment.GetHTMLURL(),
 		Color:       c.ColorScheme.Color(IssueCommented, true),
 		Fields:      fields,
@@ -250,7 +269,7 @@ func (c *Config) makeIssueCommentEmbed(ev *github.IssueCommentEvent) discord.Emb
 func (c *Config) makeIssueCommentDeletedEmbed(ev *github.IssueCommentEvent) discord.Embed {
 	return discord.Embed{
 		Title:       fmt.Sprintf("Deleted comment on issue #%d", ev.GetIssue().GetNumber()),
-		Description: markdown.Convert(ev.GetComment().GetBody(), ""),
+		Description: convertMarkdown(ev.GetComment().GetBody(), ""),
 		Author: &discord.EmbedAuthor{
 			URL:  ev.GetSender().GetHTMLURL(),
 			Name: ev.GetSender().GetLogin(),
@@ -272,28 +291,28 @@ func (c *Config) makePREmbed(ev *github.PullRequestEvent) discord.Embed {
 	if len(pr.Labels) > 0 {
 		fields = append(fields, discord.EmbedField{
 			Name:  "Labels",
-			Value: markdown.ConvertLabels(pr.Labels),
+			Value: convertLabels(pr.Labels),
 		})
 	}
 
 	if len(pr.Assignees) > 0 {
 		fields = append(fields, discord.EmbedField{
 			Name:  "Assignees",
-			Value: markdown.ConvertUsers(pr.Assignees),
+			Value: convertUsers(pr.Assignees),
 		})
 	}
 
 	if len(pr.RequestedReviewers) > 0 {
 		fields = append(fields, discord.EmbedField{
 			Name:  "Requested reviewers",
-			Value: markdown.ConvertUsers(pr.RequestedReviewers),
+			Value: convertUsers(pr.RequestedReviewers),
 		})
 	}
 
 	if len(pr.RequestedTeams) > 0 {
 		fields = append(fields, discord.EmbedField{
 			Name:  "Requested teams",
-			Value: markdown.ConvertTeams(pr.RequestedTeams),
+			Value: convertTeams(pr.RequestedTeams),
 		})
 	}
 
@@ -313,7 +332,7 @@ func (c *Config) makePREmbed(ev *github.PullRequestEvent) discord.Embed {
 			Name: pr.GetUser().GetLogin(),
 			Icon: pr.GetUser().GetAvatarURL(),
 		},
-		Description: markdown.Convert(pr.GetBody(), pr.GetHTMLURL()),
+		Description: convertMarkdown(pr.GetBody(), pr.GetHTMLURL()),
 		Color:       c.ColorScheme.Color(IssueOpened, true),
 		Fields:      fields,
 	}
@@ -522,7 +541,7 @@ func (c *Config) makePRReviewEmbed(ev *github.PullRequestReviewEvent) discord.Em
 
 	return discord.Embed{
 		Title:       fmt.Sprintf("Review submitted on pull request #%d", pr.GetNumber()),
-		Description: markdown.Convert(review.GetBody(), review.GetHTMLURL()),
+		Description: convertMarkdown(review.GetBody(), review.GetHTMLURL()),
 		URL:         review.GetHTMLURL(),
 		Color:       c.ColorScheme.Color(Reviewed, true),
 		Author: &discord.EmbedAuthor{
@@ -540,7 +559,7 @@ func (c *Config) makePRReviewDismissedEmbed(ev *github.PullRequestReviewEvent) d
 
 	return discord.Embed{
 		Title:       fmt.Sprintf("Review dismissed on pull request #%d", pr.GetNumber()),
-		Description: markdown.Convert(review.GetBody(), review.GetHTMLURL()),
+		Description: convertMarkdown(review.GetBody(), review.GetHTMLURL()),
 		URL:         review.GetHTMLURL(),
 		Color:       c.ColorScheme.Color(ReviewDismissed, true),
 		Author: &discord.EmbedAuthor{
@@ -570,7 +589,7 @@ func (c *Config) makePRReviewCommentEmbed(ev *github.PullRequestReviewCommentEve
 
 	return discord.Embed{
 		Title:       fmt.Sprintf("Review comment on pull request #%d", pr.GetNumber()),
-		Description: markdown.Convert(comment.GetBody(), comment.GetHTMLURL()),
+		Description: convertMarkdown(comment.GetBody(), comment.GetHTMLURL()),
 		URL:         comment.GetHTMLURL(),
 		Color:       c.ColorScheme.Color(ReviewCommented, true),
 		Fields:      fields,
@@ -609,14 +628,14 @@ func (c *Config) makePRReviewThreadEmbed(ev *github.PullRequestReviewThreadEvent
 	for _, comment := range t.Comments {
 		fields = append(fields, discord.EmbedField{
 			Name:   fmt.Sprintf("Review comment %d", comment.GetID()),
-			Value:  markdown.Convert(comment.GetBody(), comment.GetHTMLURL()),
+			Value:  convertMarkdown(comment.GetBody(), comment.GetHTMLURL()),
 			Inline: false,
 		})
 	}
 
 	return discord.Embed{
 		Title:  fmt.Sprintf("Review thread on pull request #%d", pr.GetNumber()),
-		URL:    markdown.ConvertThreadURL(t),
+		URL:    threadURL(t),
 		Fields: fields,
 		Color:  c.ColorScheme.Color(ReviewThreaded, true),
 		Author: &discord.EmbedAuthor{
@@ -632,7 +651,7 @@ func (c *Config) makePRReviewThreadResolvedEmbed(ev *github.PullRequestReviewThr
 
 	return discord.Embed{
 		Title: fmt.Sprintf("Review thread resolved on pull request #%d", pr.GetNumber()),
-		URL:   markdown.ConvertThreadURL(t),
+		URL:   threadURL(t),
 		Color: c.ColorScheme.Color(ReviewThreadResolved, true),
 		Author: &discord.EmbedAuthor{
 			URL:  ev.GetSender().GetHTMLURL(),
@@ -647,7 +666,7 @@ func (c *Config) makePRReviewThreadUnresolvedEmbed(ev *github.PullRequestReviewT
 
 	return discord.Embed{
 		Title: fmt.Sprintf("Review thread unresolved on pull request #%d", pr.GetNumber()),
-		URL:   markdown.ConvertThreadURL(t),
+		URL:   threadURL(t),
 		Color: c.ColorScheme.Color(ReviewThreadUnresolved, true),
 		Author: &discord.EmbedAuthor{
 			URL:  ev.GetSender().GetHTMLURL(),
@@ -658,6 +677,92 @@ func (c *Config) makePRReviewThreadUnresolvedEmbed(ev *github.PullRequestReviewT
 }
 
 /// END PullRequestReviewThreadEvent Discord embeds
+
+func convertSlicePtr[T any](s []*T) []T {
+	r := make([]T, 0, len(s))
+	for _, v := range s {
+		r = append(r, *v)
+	}
+	return r
+}
+
+func convertLabels[T *github.Label | github.Label](labelsV []T) string {
+	var labels []github.Label
+	switch labelsV := any(labelsV).(type) {
+	case []github.Label:
+		labels = labelsV
+	case []*github.Label:
+		labels = convertSlicePtr(labelsV)
+	}
+
+	var labelNames links
+	for _, label := range labels {
+		labelNames = append(labelNames, link{
+			title: label.GetName(),
+			href:  label.GetURL(),
+		})
+	}
+
+	return labelNames.commaSeparatedList()
+}
+
+func threadURL(t *github.PullRequestThread) (url string) {
+	comment := t.Comments[0]
+	if comment != nil {
+		url = comment.GetHTMLURL()
+	}
+	return
+}
+
+func convertUsers(users []*github.User) string {
+	var names links
+	for _, user := range users {
+		names = append(names, link{title: user.GetLogin(), href: user.GetHTMLURL()})
+	}
+	return names.commaSeparatedList()
+}
+
+func convertTeams(teams []*github.Team) string {
+	var names links
+	for _, team := range teams {
+		names = append(names, link{title: team.GetName(), href: team.GetHTMLURL()})
+	}
+	return names.commaSeparatedList()
+}
+
+var mdMaxSize = 4096
+var mdParser = parser.NewParser(
+	parser.WithBlockParsers(discordmd.BlockParsers()...),
+	parser.WithInlineParsers(discordmd.InlineParserWithLink()...),
+)
+
+func convertHyperlink(content, href string) string {
+	if href == "" {
+		return content
+	}
+	return fmt.Sprintf("[%s](%s)", content, href)
+}
+
+func convertMarkdown(githubMD, readMoreURL string) string {
+	bytes, buff := []byte(githubMD), strings.Builder{}
+	node := mdParser.Parse(text.NewReader(bytes))
+	if err := discordmd.DefaultRenderer.Render(&buff, bytes, node); err != nil {
+		return githubMD
+	}
+
+	s := buff.String()
+	if len(s) <= mdMaxSize {
+		return s
+	}
+
+	// If the message is too long, we need to truncate it.
+	truncateMsg := "... (truncated)"
+	if readMoreURL != "" {
+		truncateMsg = fmt.Sprintf("... (%s)", convertHyperlink("_read more_", readMoreURL))
+	}
+
+	return s[:strings.LastIndex(s[:mdMaxSize-len(truncateMsg)], " ")] + truncateMsg
+}
 
 // checkPR checks if an issue happens to be a pull request
 func checkPR(issue *github.Issue) bool {
