+++ go.mod
@@ -3,20 +3,20 @@ module github.com/ethanthatonekid/gitcord
 go 1.19
 
 require (
-	github.com/diamondburned/ningen/v3 v3.0.0-20221109004709-fed56ccbcd19
+	github.com/diamondburned/ningen/v3 v3.0.0-20221020234758-38b078a9ab7b
 	github.com/google/go-github/v47 v47.1.0
 	github.com/yuin/goldmark v1.3.2
-	go4.org v0.0.0-20201209231011-d4a079459e60
 )
 
 require (
 	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
+	go4.org v0.0.0-20200411211856-f5505b9728dd // indirect
 )
 
 require (
-	github.com/diamondburned/arikawa/v3 v3.1.1-0.20221103093025-87c479a2dcd4
+	github.com/diamondburned/arikawa/v3 v3.1.1-0.20221014060129-181dcb1bdd00
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/google/go-querystring v1.1.0 // indirect
 	github.com/gorilla/schema v1.2.0 // indirect
